requirements.txt
streamlit==1.50.0
pillow
numpy
opencv-python
